requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
ephem==3.7.7.1
PySocks==1.7.1
python-telegram-bot==13.5
pytz==2021.1
six==1.16.0
tornado==6.1
tzlocal==2.1
emoji==1.2.0
clarifai==2.6.2
googleapis-common-protos==1.53.0
grpcio==1.36.1
protobuf==3.15.6
requests==2.25.1
